# New AsyncWriteSortingCollection