--- a/src/main/java/htsjdk/samtools/util/SortingCollection.java
+++ b/src/main/java/htsjdk/samtools/util/SortingCollection.java
@@ -34,14 +34,14 @@ import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.Serializable;
 import java.lang.reflect.Array;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Comparator;
 import java.util.Iterator;
-import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Queue;
 import java.util.TreeSet;
+import java.util.concurrent.ConcurrentLinkedQueue;
 
 /**
  * Collection to which many records can be added.  After all records are added, the collection can be
@@ -96,39 +96,39 @@ public class SortingCollection<T> implements Iterable<T> {
     }
 
     /** Directories where files of sorted records go. */
-    private final File[] tmpDirs;
+    protected final File[] tmpDirs;
 
     /** The minimum amount of space free on a temp filesystem to write a file there. */
-    private final long TMP_SPACE_FREE = IOUtil.FIVE_GBS;
+    protected final long TMP_SPACE_FREE = IOUtil.FIVE_GBS;
 
     /**
      * Used to write records to file, and used as a prototype to create codecs for reading.
      */
-    private final SortingCollection.Codec<T> codec;
+    protected final SortingCollection.Codec<T> codec;
 
     /**
      * For sorting, both when spilling records to file, and merge sorting.
      */
-    private final Comparator<T> comparator;
-    private final int maxRecordsInRam;
-    private int numRecordsInRam = 0;
-    private T[] ramRecords;
-    private boolean iterationStarted = false;
-    private boolean doneAdding = false;
+    protected final Comparator<T> comparator;
+    protected final int maxRecordsInRam;
+    protected int numRecordsInRam = 0;
+    protected T[] ramRecords;
+    protected boolean iterationStarted = false;
+    protected boolean doneAdding = false;
 
     /**
      * Set to true when all temp files have been cleaned up
      */
-    private boolean cleanedUp = false;
+    protected boolean cleanedUp = false;
 
     /**
      * List of files in tmpDir containing sorted records
      */
-    private final List<File> files = new ArrayList<File>();
+    protected final Queue<File> files = new ConcurrentLinkedQueue<>();
 
-    private boolean destructiveIteration = true;
+    protected boolean destructiveIteration = true;
 
-    private TempStreamFactory tempStreamFactory = new TempStreamFactory();
+    protected TempStreamFactory tempStreamFactory = new TempStreamFactory();
 
     /**
      * Prepare to accumulate records to be sorted
@@ -138,7 +138,7 @@ public class SortingCollection<T> implements Iterable<T> {
      * @param maxRecordsInRam how many records to accumulate before spilling to disk
      * @param tmpDir Where to write files of records that will not fit in RAM
      */
-    private SortingCollection(final Class<T> componentType, final SortingCollection.Codec<T> codec,
+    protected SortingCollection(final Class<T> componentType, final SortingCollection.Codec<T> codec,
                              final Comparator<T> comparator, final int maxRecordsInRam, final File... tmpDir) {
         if (maxRecordsInRam <= 0) {
             throw new IllegalArgumentException("maxRecordsInRam must be > 0");
@@ -163,9 +163,13 @@ public class SortingCollection<T> implements Iterable<T> {
             throw new IllegalStateException("Cannot add after calling iterator()");
         }
         if (numRecordsInRam == maxRecordsInRam) {
-            spillToDisk();
+            performSpillToDisk();
         }
         ramRecords[numRecordsInRam++] = rec;
+    }
+
+    void performSpillToDisk() {
+        defaultSpillToDisk();
     }
 
     /**
@@ -183,16 +187,22 @@ public class SortingCollection<T> implements Iterable<T> {
 
         doneAdding = true;
 
+        finish();
+
         if (this.files.isEmpty()) {
             return;
         }
 
         if (this.numRecordsInRam > 0) {
-            spillToDisk();
+            defaultSpillToDisk();
         }
 
         // Facilitate GC
         this.ramRecords = null;
+    }
+
+    void finish() {
+        // hook to be overriden, do nothing here
     }
 
     /**
@@ -214,18 +224,24 @@ public class SortingCollection<T> implements Iterable<T> {
     /**
      * Sort the records in memory, write them to a file, and clear the buffer of records in memory.
      */
-    private void spillToDisk() {
+    protected void defaultSpillToDisk() {
+        Arrays.sort(this.ramRecords, 0, this.numRecordsInRam, this.comparator);
+        spill(this.ramRecords, this.numRecordsInRam, this.codec);
+        numRecordsInRam = 0;
+    }
+
+    void spill(T[] buffRamRecords, int buffNumRecordsInRam, Codec<T> codec) {
         try {
-            Arrays.sort(this.ramRecords, 0, this.numRecordsInRam, this.comparator);
             final File f = newTempFile();
             OutputStream os = null;
             try {
-                os = tempStreamFactory.wrapTempOutputStream(new FileOutputStream(f), Defaults.BUFFER_SIZE);
-                this.codec.setOutputStream(os);
-                for (int i = 0; i < this.numRecordsInRam; ++i) {
-                    this.codec.encode(ramRecords[i]);
+                os = tempStreamFactory.wrapTempOutputStream(
+                        new FileOutputStream(f), Defaults.BUFFER_SIZE);
+                codec.setOutputStream(os);
+                for (int i = 0; i < buffNumRecordsInRam; ++i) {
+                    codec.encode(buffRamRecords[i]);
                     // Facilitate GC
-                    this.ramRecords[i] = null;
+                    buffRamRecords[i] = null;
                 }
 
                 os.flush();
@@ -238,7 +254,6 @@ public class SortingCollection<T> implements Iterable<T> {
                 }
             }
 
-            this.numRecordsInRam = 0;
             this.files.add(f);
 
         }
@@ -298,8 +313,13 @@ public class SortingCollection<T> implements Iterable<T> {
                                                        final Comparator<T> comparator,
                                                        final int maxRecordsInRAM,
                                                        final File... tmpDir) {
-        return new SortingCollection<T>(componentType, codec, comparator, maxRecordsInRAM, tmpDir);
-
+        if (Defaults.SORTING_COLLECTION_THREADS > 0) {
+            return new AsyncWriteSortingCollection<>(componentType, codec,
+                    comparator, maxRecordsInRAM, tmpDir);
+        } else {
+            return new SortingCollection<>(componentType, codec, comparator,
+                    maxRecordsInRAM, tmpDir);
+        }
     }
 
     /**
@@ -316,12 +336,13 @@ public class SortingCollection<T> implements Iterable<T> {
                                                        final Comparator<T> comparator,
                                                        final int maxRecordsInRAM,
                                                        final Collection<File> tmpDirs) {
-        return new SortingCollection<T>(componentType,
-                                        codec,
-                                        comparator,
-                                        maxRecordsInRAM,
-                                        tmpDirs.toArray(new File[tmpDirs.size()]));
-
+        if (Defaults.SORTING_COLLECTION_THREADS > 0) {
+            return new AsyncWriteSortingCollection<>(componentType, codec,
+                    comparator, maxRecordsInRAM, tmpDirs.toArray(new File[tmpDirs.size()]));
+        } else {
+            return new SortingCollection<>(componentType, codec, comparator,
+                    maxRecordsInRAM, tmpDirs.toArray(new File[tmpDirs.size()]));
+        }
     }
 
 
@@ -339,7 +360,13 @@ public class SortingCollection<T> implements Iterable<T> {
                                                        final int maxRecordsInRAM) {
 
         final File tmpDir = new File(System.getProperty("java.io.tmpdir"));
-        return new SortingCollection<T>(componentType, codec, comparator, maxRecordsInRAM, tmpDir);
+        if (Defaults.SORTING_COLLECTION_THREADS > 0) {
+            return new AsyncWriteSortingCollection<>(componentType, codec,
+                    comparator, maxRecordsInRAM, tmpDir);
+        } else {
+            return new SortingCollection<>(componentType, codec, comparator,
+                    maxRecordsInRAM, tmpDir);
+        }
     }
 
     /**