--- a/src/test/java/htsjdk/samtools/util/SortingCollectionTest.java
+++ b/src/test/java/htsjdk/samtools/util/SortingCollectionTest.java
@@ -90,9 +90,10 @@ public class SortingCollectionTest extends HtsjdkTest {
             sortingCollection.add(s);
             strings[numStringsGenerated++] = s;
         }
+        sortingCollection.doneAdding();
         Arrays.sort(strings, new StringComparator());
 
-        Assert.assertEquals(tmpDirIsEmpty(), numStringsToGenerate <= maxRecordsInRam);
+        Assert.assertEquals(tmpDirIsEmpty(), shouldTmpDirBeEmpty(numStringsToGenerate, maxRecordsInRam));
         sortingCollection.setDestructiveIteration(false);
         assertIteratorEqualsList(strings, sortingCollection.iterator());
         assertIteratorEqualsList(strings, sortingCollection.iterator());
@@ -101,7 +102,7 @@ public class SortingCollectionTest extends HtsjdkTest {
         Assert.assertEquals(tmpDir().list().length, 0);
     }
 
-    private void assertIteratorEqualsList(final String[] strings, final Iterator<String> sortingCollection) {
+    protected void assertIteratorEqualsList(final String[] strings, final Iterator<String> sortingCollection) {
         int i = 0;
         while (sortingCollection.hasNext()) {
             final String s = sortingCollection.next();
@@ -110,8 +111,13 @@ public class SortingCollectionTest extends HtsjdkTest {
         Assert.assertEquals(i, strings.length);
     }
 
-    private SortingCollection<String> makeSortingCollection(final int maxRecordsInRam) {
-        return SortingCollection.newInstance(String.class, new StringCodec(), new StringComparator(), maxRecordsInRam, tmpDir());
+    boolean shouldTmpDirBeEmpty(int numStringsToGenerate, int maxRecordsInRam) {
+        return numStringsToGenerate <= maxRecordsInRam;
+    }
+
+    SortingCollection<String> makeSortingCollection(final int maxRecordsInRam) {
+            return new SortingCollection<>(String.class, new StringCodec(),
+                    new StringComparator(), maxRecordsInRam, tmpDir());
     }
 
     /**