We present a multithreaded implementation of SortingCollection.
The main idea of this improvements is that data is being sorted and spilled in the sorting collection asynchronously by submitting tasks to the executor service.
Compared to the current implementation it looks as follows:

![compare](https://user-images.githubusercontent.com/11179595/27478925-52b3e8ec-581a-11e7-87aa-ba1421d3195f.png)

As we can see above, parallel version enables continuous sorting work without breaks for writing to the temporary file.
Method doneAdding() is blocking, since we have to wait until all spill tasks are completed. 

What is the effect?
Below are results of the SortingCollection benchmark. It is based on the random generation of data for SortingCollection, we just create random int, transform it to string and put into the sorting collection.

![image](https://user-images.githubusercontent.com/11179595/27478961-6c59bfe2-581a-11e7-9384-778dc008b45b.png)

```
Benchmark                                              Number of strings  Mode  Cnt      Score      Error  Units
SortingCollectionBenchmark.sortingCollectionBenchmark         50_000_000  avgt    5  42821.407 ± 2095.994  ms/op
SortingCollectionBenchmark.sortingCollectionBenchmark2th      50_000_000  avgt    5  19241.607 ±  353.917  ms/op
SortingCollectionBenchmark.sortingCollectionBenchmark4th      50_000_000  avgt    5  10750.524 ± 72.900    ms/op
```

Here is log output of SortSam metric for both SortingCollection implementattions:
```
AsyncSpillSortingCollection with 2 spilling threads:
INFO 2017-06-14 15:58:47 SortSam Read 10,000,000 records. Elapsed time: 00:00:22s. Time for last 10,000,000: 22s. Last read position: 1:148,346,706
INFO 2017-06-14 15:59:14 SortSam Read 20,000,000 records. Elapsed time: 00:00:49s. Time for last 10,000,000: 26s. Last read position: X:149,840,001
INFO 2017-06-14 15:59:41 SortSam Read 30,000,000 records. Elapsed time: 00:01:16s. Time for last 10,000,000: 27s. Last read position: 12:31,820,558
INFO 2017-06-14 16:00:09 SortSam Read 40,000,000 records. Elapsed time: 00:01:44s. Time for last 10,000,000: 27s. Last read position: 11:3,746,487
INFO 2017-06-14 16:00:36 SortSam Read 50,000,000 records. Elapsed time: 00:02:11s. Time for last 10,000,000: 27s. Last read position: 17:27,904,169
INFO 2017-06-14 16:00:47 SortSam Finished reading inputs, merging and writing to output now.
Total elapsed time = 129 sec - performance gain is 72%
```
```
Standard SortingCollection:
INFO 2017-06-14 16:05:36 SortSam Read 10,000,000 records. Elapsed time: 00:00:40s. Time for last 10,000,000: 40s. Last read position: 1:148,346,706
INFO 2017-06-14 16:06:20 SortSam Read 20,000,000 records. Elapsed time: 00:01:25s. Time for last 10,000,000: 44s. Last read position: X:149,840,001
INFO 2017-06-14 16:07:05 SortSam Read 30,000,000 records. Elapsed time: 00:02:10s. Time for last 10,000,000: 45s. Last read position: 12:31,820,558
INFO 2017-06-14 16:07:51 SortSam Read 40,000,000 records. Elapsed time: 00:02:55s. Time for last 10,000,000: 45s. Last read position: 11:3,746,487
INFO 2017-06-14 16:08:40 SortSam Read 50,000,000 records. Elapsed time: 00:03:44s. Time for last 10,000,000: 48s. Last read position: 17:27,904,169
INFO 2017-06-14 16:08:58 SortSam Finished reading inputs, merging and writing to output now.
       Total elapsed time = 222 sec 
```

```
Command for run:
standard version: java -Xmx4g -jar picard.jar SortSam I=/path/to/bam O=/path/to/bam SORT_ORDER=queryname VALIDATION_STRINGENCY=SILENT
parallel version: java -Dsamjdk.sort_col_threads=2 -Xmx4g -jar picard-unspecified-SNAPSHOT-all.jar SortSam I=/path/to/bam O=/path/to/out SORT_ORDER=queryname VALIDATION_STRINGENCY=SILENT
```
The parallel processing is turned ON by setting JVM option  -Dsamjdk.sort_col_threads=*number_of_threads*, if *number_of_threads* > 0, any metric/utility that uses SortingCollection will use the parallel implementation.