@tfenne 2017-06-23
@SilinPavel Thanks for doing this.  I'll do a full review later but wanted to share some thoughts up front:

1. I'm surprised at the gains you see in SortSam.  I looked into this a while back and it seemed to me that so much time was spent in the GC with all those objects in memory that parallelizing wouldn't help.  In fact, I'm concerned that the speedup you observe has more to do with the shrinking of the number of records in RAM (I see it's set to `MAX / (threads+1)`) than the parallelism.  In my own testing just now, simply running the existing SortSam with `MAX_RECORDS_IN_RAM=166666` vs. `MAX_RECORDS_IN_RAM=500000` gave me a performance gain almost identical to that you see in SortSam.  Your test shows `129/222 = 0.581`, mine on 20m reads was `82/129 = 0.635`.
2. I'm also concerned about the up-tick in the number of files used with this approach.  I.e. if I'm sorting a file with 1bn records and set `MAX_RECORDS_IN_RAM=1000000`, I'd expect it to use 1000 files, which is within most systems limit for open filehandles, but if set to use two threads it'll create 3000 files which will cause problems for many people.
3. I've been thinking about parallelism in htsjdk in general and am keen to introduce it in a way that can easily be re-used across multiple pieces of the code, e.g. BAM block compression.  As a user I think what we want is to be able to set the number of threads available for _all_ parallel work, not per object using parallelism.  E.g. in Picard's IlluminaBasecallsToSam there are multiple sorting collections being used in parallel - it would be great to allow say 8 threads used it total regardless of the total number of sorting collections.  To do this I was thinking that we would create a singleton object in htsjdk that holds a single `ExecutorService` and has methods to allow users to set/override the `ExecutorService` in use.  The service could default to a subclass of `AbstractExecutorService` that performs synchronous execution, so that the default behaviour is effectively serial.  Then if the number of threads is set to > 0 it could be substituted with a `ForkJoinPool` or similar.  This would obviously change what you've done a fair bit, including e.g. renaming the parameter/Defaults entry to be something like `htsjdk.parallel_threads`.
4. If (3) were done, I think the new AsyncSortingCollection could be reduced to a fairly small set of changes in SortingCollection instead of a new class.
5. Tests are going to be very important for this.  I would suggest that `SortingCollectionTest` would need a) to have all test cases run for serial and parallel and b) new tests added to ensure that failures in sorting/writing are propagated correctly in parallel mode.