@SilinPavel 2017-06-28
@tfenne thanks for your response.
Here are our answers:
1. Performance gain is not a surprise for us, since:
A. We witness it in one of our customers' pipelines running in AWS, were we are evaluating this updated SortingCollection and actually this pull request has a clear goal A - to satisfy that customer and let him get the important results faster.
B. the proposed changes lead to distribution of work between two independent process: outside work with SortingCollection (f.i. add()) and inside processes such as sorting of a buffer and writing it to disk. 
So performance gain seems quite natural. Below are some tests that should prove it. Also we are attaching GC jvm logs and we can see that in our case GC activity is not so high and affects on performance not so much as you mentioned before. 

```
java -XX:+UnlockCommercialFeatures -XX:+FlightRecorder -XX:StartFlightRecording=dumponexit=true,filename=sc.jfr -jar picard-2.7.1-87-g7a85c95-SNAPSHOT-all.jar SortSam I=/home/pavel_silin/genomic_data/bams/NA12878.mapped.ILLUMINA.bwa.CEU.exome.20121211.bam O=/home/pavel_silin/sort_sc.bam SORT_ORDER=queryname VALIDATION_STRINGENCY=SILENT
 
 
[Mon Jun 26 16:57:34 MSK 2017] picard.sam.SortSam INPUT=/home/pavel_silin/genomic_data/bams/NA12878.mapped.ILLUMINA.bwa.CEU.exome.20121211.bam OUTPUT=/home/pavel_silin/sort_sc.bam SORT_ORDER=queryname VALIDATION_STRINGENCY=SILENT    VERBOSITY=INFO QUIET=false COMPRESSION_LEVEL=5 MAX_RECORDS_IN_RAM=500000 CREATE_INDEX=false CREATE_MD5_FILE=false GA4GH_CLIENT_SECRETS=client_secrets.json
[Mon Jun 26 16:57:34 MSK 2017] Executing as pavel_silin@EPRUPETW6431 on Linux 4.4.0-81-generic amd64; Java HotSpot(TM) 64-Bit Server VM 1.8.0_131-b11; Picard version: 2.7.1-87-g7a85c95-SNAPSHOT
INFO    2017-06-26 16:58:24 SortSam Read    10,000,000 records.  Elapsed time: 00:00:49s.  Time for last 10,000,000:   49s.  Last read position: 1:119,650,743
INFO    2017-06-26 16:59:14 SortSam Read    20,000,000 records.  Elapsed time: 00:01:39s.  Time for last 10,000,000:   49s.  Last read position: 1:248,374,319
INFO    2017-06-26 17:00:07 SortSam Read    30,000,000 records.  Elapsed time: 00:02:32s.  Time for last 10,000,000:   53s.  Last read position: 2:154,501,231
INFO    2017-06-26 17:01:00 SortSam Read    40,000,000 records.  Elapsed time: 00:03:25s.  Time for last 10,000,000:   52s.  Last read position: 3:49,897,096
INFO    2017-06-26 17:01:49 SortSam Read    50,000,000 records.  Elapsed time: 00:04:14s.  Time for last 10,000,000:   49s.  Last read position: 4:9,702,042
INFO    2017-06-26 17:02:39 SortSam Read    60,000,000 records.  Elapsed time: 00:05:04s.  Time for last 10,000,000:   49s.  Last read position: 5:9,019,531
INFO    2017-06-26 17:03:28 SortSam Read    70,000,000 records.  Elapsed time: 00:05:53s.  Time for last 10,000,000:   49s.  Last read position: 5:176,089,346
INFO    2017-06-26 17:04:18 SortSam Read    80,000,000 records.  Elapsed time: 00:06:43s.  Time for last 10,000,000:   49s.  Last read position: 6:143,265,238
INFO    2017-06-26 17:05:07 SortSam Read    90,000,000 records.  Elapsed time: 00:07:32s.  Time for last 10,000,000:   49s.  Last read position: 7:127,343,317
INFO    2017-06-26 17:05:56 SortSam Read   100,000,000 records.  Elapsed time: 00:08:22s.  Time for last 10,000,000:   49s.  Last read position: 8:133,732,360
INFO    2017-06-26 17:06:46 SortSam Read   110,000,000 records.  Elapsed time: 00:09:11s.  Time for last 10,000,000:   49s.  Last read position: 10:3,118,522
INFO    2017-06-26 17:07:35 SortSam Read   120,000,000 records.  Elapsed time: 00:10:01s.  Time for last 10,000,000:   49s.  Last read position: 11:9,164,230
INFO    2017-06-26 17:08:24 SortSam Read   130,000,000 records.  Elapsed time: 00:10:50s.  Time for last 10,000,000:   48s.  Last read position: 12:6,800,342
INFO    2017-06-26 17:09:14 SortSam Read   140,000,000 records.  Elapsed time: 00:11:39s.  Time for last 10,000,000:   49s.  Last read position: 13:25,887,091
INFO    2017-06-26 17:10:03 SortSam Read   150,000,000 records.  Elapsed time: 00:12:28s.  Time for last 10,000,000:   49s.  Last read position: 14:94,527,360
INFO    2017-06-26 17:10:52 SortSam Read   160,000,000 records.  Elapsed time: 00:13:18s.  Time for last 10,000,000:   49s.  Last read position: 16:20,823,053
INFO    2017-06-26 17:11:42 SortSam Read   170,000,000 records.  Elapsed time: 00:14:07s.  Time for last 10,000,000:   49s.  Last read position: 17:40,003,660
INFO    2017-06-26 17:12:31 SortSam Read   180,000,000 records.  Elapsed time: 00:14:56s.  Time for last 10,000,000:   49s.  Last read position: 19:11,138,565
INFO    2017-06-26 17:13:20 SortSam Read   190,000,000 records.  Elapsed time: 00:15:46s.  Time for last 10,000,000:   49s.  Last read position: 20:57,487,759
INFO    2017-06-26 17:14:10 SortSam Read   200,000,000 records.  Elapsed time: 00:16:35s.  Time for last 10,000,000:   49s.  Last read position: X:57,813,555
INFO    2017-06-26 17:15:01 SortSam Read   210,000,000 records.  Elapsed time: 00:17:26s.  Time for last 10,000,000:   51s.  Last read position: hs37d5:19,940,511
INFO    2017-06-26 17:15:14 SortSam Finished reading inputs, merging and writing to output now.
 
Elapsed time: 1059 secs
```
   
 | Young Collection Total Time | Old Collection Total Time | All Collections Total Time
 | ------------------------------------- | --------------------------------- | ----------------------------------
GC Count | 237 | 2 | 239
Average GC Time | 45 ms | 115 ms | 46 ms
Maximum GC Time | 194 ms | 128 ms | 194 ms
Total GC Time| 10 s | 230 ms | 11 s

```
java -XX:+UnlockCommercialFeatures -XX:+FlightRecorder -XX:StartFlightRecording=dumponexit=true,filename=sc_less_max+record.jfr -jar picard-2.7.1-87-g7a85c95-SNAPSHOT-all.jar SortSam I=/home/pavel_silin/genomic_data/bams/NA12878.mapped.ILLUMINA.bwa.CEU.exome.20121211.bam O=/home/pavel_silin/sort_sc.bam SORT_ORDER=queryname VALIDATION_STRINGENCY=SILENT MAX_RECORDS_IN_RAM=166666
 
 
[Mon Jun 26 19:37:33 MSK 2017] picard.sam.SortSam INPUT=/home/pavel_silin/genomic_data/bams/NA12878.mapped.ILLUMINA.bwa.CEU.exome.20121211.bam OUTPUT=/home/pavel_silin/sort_sc.bam SORT_ORDER=queryname VALIDATION_STRINGENCY=SILENT MAX_RECORDS_IN_RAM=166666    VERBOSITY=INFO QUIET=false COMPRESSION_LEVEL=5 CREATE_INDEX=false CREATE_MD5_FILE=false GA4GH_CLIENT_SECRETS=client_secrets.json
[Mon Jun 26 19:37:33 MSK 2017] Executing as pavel_silin@EPRUPETW6431 on Linux 4.4.0-81-generic amd64; Java HotSpot(TM) 64-Bit Server VM 1.8.0_131-b11; Picard version: 2.7.1-87-g7a85c95-SNAPSHOT
INFO    2017-06-26 19:38:23 SortSam Read    10,000,000 records.  Elapsed time: 00:00:49s.  Time for last 10,000,000:   49s.  Last read position: 1:119,650,743
INFO    2017-06-26 19:39:14 SortSam Read    20,000,000 records.  Elapsed time: 00:01:40s.  Time for last 10,000,000:   50s.  Last read position: 1:248,374,319
INFO    2017-06-26 19:40:03 SortSam Read    30,000,000 records.  Elapsed time: 00:02:29s.  Time for last 10,000,000:   48s.  Last read position: 2:154,501,231
INFO    2017-06-26 19:40:52 SortSam Read    40,000,000 records.  Elapsed time: 00:03:17s.  Time for last 10,000,000:   48s.  Last read position: 3:49,897,096
INFO    2017-06-26 19:41:41 SortSam Read    50,000,000 records.  Elapsed time: 00:04:06s.  Time for last 10,000,000:   49s.  Last read position: 4:9,702,042
INFO    2017-06-26 19:42:30 SortSam Read    60,000,000 records.  Elapsed time: 00:04:55s.  Time for last 10,000,000:   48s.  Last read position: 5:9,019,531
INFO    2017-06-26 19:43:18 SortSam Read    70,000,000 records.  Elapsed time: 00:05:44s.  Time for last 10,000,000:   48s.  Last read position: 5:176,089,346
INFO    2017-06-26 19:44:07 SortSam Read    80,000,000 records.  Elapsed time: 00:06:33s.  Time for last 10,000,000:   48s.  Last read position: 6:143,265,238
INFO    2017-06-26 19:44:55 SortSam Read    90,000,000 records.  Elapsed time: 00:07:21s.  Time for last 10,000,000:   48s.  Last read position: 7:127,343,317
INFO    2017-06-26 19:45:44 SortSam Read   100,000,000 records.  Elapsed time: 00:08:10s.  Time for last 10,000,000:   48s.  Last read position: 8:133,732,360
INFO    2017-06-26 19:46:32 SortSam Read   110,000,000 records.  Elapsed time: 00:08:58s.  Time for last 10,000,000:   47s.  Last read position: 10:3,118,522
INFO    2017-06-26 19:47:20 SortSam Read   120,000,000 records.  Elapsed time: 00:09:46s.  Time for last 10,000,000:   48s.  Last read position: 11:9,164,230
INFO    2017-06-26 19:48:08 SortSam Read   130,000,000 records.  Elapsed time: 00:10:34s.  Time for last 10,000,000:   48s.  Last read position: 12:6,800,342
INFO    2017-06-26 19:48:56 SortSam Read   140,000,000 records.  Elapsed time: 00:11:22s.  Time for last 10,000,000:   48s.  Last read position: 13:25,887,091
INFO    2017-06-26 19:49:45 SortSam Read   150,000,000 records.  Elapsed time: 00:12:11s.  Time for last 10,000,000:   48s.  Last read position: 14:94,527,360
INFO    2017-06-26 19:50:33 SortSam Read   160,000,000 records.  Elapsed time: 00:12:59s.  Time for last 10,000,000:   48s.  Last read position: 16:20,823,053
INFO    2017-06-26 19:51:21 SortSam Read   170,000,000 records.  Elapsed time: 00:13:47s.  Time for last 10,000,000:   48s.  Last read position: 17:40,003,660
INFO    2017-06-26 19:52:09 SortSam Read   180,000,000 records.  Elapsed time: 00:14:35s.  Time for last 10,000,000:   48s.  Last read position: 19:11,138,565
INFO    2017-06-26 19:52:57 SortSam Read   190,000,000 records.  Elapsed time: 00:15:22s.  Time for last 10,000,000:   47s.  Last read position: 20:57,487,759
INFO    2017-06-26 19:53:45 SortSam Read   200,000,000 records.  Elapsed time: 00:16:11s.  Time for last 10,000,000:   48s.  Last read position: X:57,813,555
INFO    2017-06-26 19:54:35 SortSam Read   210,000,000 records.  Elapsed time: 00:17:00s.  Time for last 10,000,000:   49s.  Last read position: hs37d5:19,940,511
INFO    2017-06-26 19:54:47 SortSam Finished reading inputs, merging and writing to output now.
 
Elapsed time: 1032 secs
```

 | Young Collection Total Time | Old Collection Total Time | All Collections Total Time
 | ------------------------------------- | --------------------------------- | ----------------------------------
GC Count | 538 | 1 | 539
Average GC Time | 15 ms | 64 ms | 15 ms
Maximum GC Time | 55 ms | 55 ms | 64 ms
Total GC Time| 8 s | 64 ms | 8 s

```
java -Dsamjdk.sort_col_threads=2 -XX:+UnlockCommercialFeatures -XX:+FlightRecorder -XX:StartFlightRecording=dumponexit=true,filename=awsc.jfr -jar picard-2.7.1-104-g4cf8f75-SNAPSHOT-all.jar SortSam I=/home/pavel_silin/genomic_data/bams/NA12878.mapped.ILLUMINA.bwa.CEU.exome.20121211.bam O=/home/pavel_silin/sort_sc.bam SORT_ORDER=queryname VALIDATION_STRINGENCY=SILENT
 
 
[Mon Jun 26 17:55:39 MSK 2017] picard.sam.SortSam INPUT=/home/pavel_silin/genomic_data/bams/NA12878.mapped.ILLUMINA.bwa.CEU.exome.20121211.bam OUTPUT=/home/pavel_silin/sort_sc.bam SORT_ORDER=queryname VALIDATION_STRINGENCY=SILENT    VERBOSITY=INFO QUIET=false COMPRESSION_LEVEL=5 MAX_RECORDS_IN_RAM=500000 CREATE_INDEX=false CREATE_MD5_FILE=false GA4GH_CLIENT_SECRETS=client_secrets.json
[Mon Jun 26 17:55:39 MSK 2017] Executing as pavel_silin@EPRUPETW6431 on Linux 4.4.0-81-generic amd64; Java HotSpot(TM) 64-Bit Server VM 1.8.0_131-b11; Picard version: 2.7.1-104-g4cf8f75-SNAPSHOT
INFO    2017-06-26 17:56:07 SortSam Read    10,000,000 records.  Elapsed time: 00:00:27s.  Time for last 10,000,000:   27s.  Last read position: 1:119,650,743
INFO    2017-06-26 17:56:33 SortSam Read    20,000,000 records.  Elapsed time: 00:00:52s.  Time for last 10,000,000:   25s.  Last read position: 1:248,374,319
INFO    2017-06-26 17:56:58 SortSam Read    30,000,000 records.  Elapsed time: 00:01:17s.  Time for last 10,000,000:   25s.  Last read position: 2:154,501,231
INFO    2017-06-26 17:57:22 SortSam Read    40,000,000 records.  Elapsed time: 00:01:41s.  Time for last 10,000,000:   24s.  Last read position: 3:49,897,096
INFO    2017-06-26 17:57:47 SortSam Read    50,000,000 records.  Elapsed time: 00:02:06s.  Time for last 10,000,000:   24s.  Last read position: 4:9,702,042
INFO    2017-06-26 17:58:12 SortSam Read    60,000,000 records.  Elapsed time: 00:02:31s.  Time for last 10,000,000:   25s.  Last read position: 5:9,019,531
INFO    2017-06-26 17:58:37 SortSam Read    70,000,000 records.  Elapsed time: 00:02:56s.  Time for last 10,000,000:   24s.  Last read position: 5:176,089,346
INFO    2017-06-26 17:59:02 SortSam Read    80,000,000 records.  Elapsed time: 00:03:22s.  Time for last 10,000,000:   25s.  Last read position: 6:143,265,238
INFO    2017-06-26 17:59:27 SortSam Read    90,000,000 records.  Elapsed time: 00:03:46s.  Time for last 10,000,000:   24s.  Last read position: 7:127,343,317
INFO    2017-06-26 17:59:52 SortSam Read   100,000,000 records.  Elapsed time: 00:04:11s.  Time for last 10,000,000:   25s.  Last read position: 8:133,732,360
INFO    2017-06-26 18:00:17 SortSam Read   110,000,000 records.  Elapsed time: 00:04:36s.  Time for last 10,000,000:   25s.  Last read position: 10:3,118,522
INFO    2017-06-26 18:00:41 SortSam Read   120,000,000 records.  Elapsed time: 00:05:00s.  Time for last 10,000,000:   24s.  Last read position: 11:9,164,230
INFO    2017-06-26 18:01:06 SortSam Read   130,000,000 records.  Elapsed time: 00:05:25s.  Time for last 10,000,000:   24s.  Last read position: 12:6,800,342
INFO    2017-06-26 18:01:30 SortSam Read   140,000,000 records.  Elapsed time: 00:05:49s.  Time for last 10,000,000:   24s.  Last read position: 13:25,887,091
INFO    2017-06-26 18:01:54 SortSam Read   150,000,000 records.  Elapsed time: 00:06:13s.  Time for last 10,000,000:   23s.  Last read position: 14:94,527,360
INFO    2017-06-26 18:02:18 SortSam Read   160,000,000 records.  Elapsed time: 00:06:38s.  Time for last 10,000,000:   24s.  Last read position: 16:20,823,053
INFO    2017-06-26 18:02:43 SortSam Read   170,000,000 records.  Elapsed time: 00:07:02s.  Time for last 10,000,000:   24s.  Last read position: 17:40,003,660
INFO    2017-06-26 18:03:07 SortSam Read   180,000,000 records.  Elapsed time: 00:07:26s.  Time for last 10,000,000:   23s.  Last read position: 19:11,138,565
INFO    2017-06-26 18:03:31 SortSam Read   190,000,000 records.  Elapsed time: 00:07:50s.  Time for last 10,000,000:   23s.  Last read position: 20:57,487,759
INFO    2017-06-26 18:03:55 SortSam Read   200,000,000 records.  Elapsed time: 00:08:14s.  Time for last 10,000,000:   24s.  Last read position: X:57,813,555
INFO    2017-06-26 18:04:20 SortSam Read   210,000,000 records.  Elapsed time: 00:08:39s.  Time for last 10,000,000:   24s.  Last read position: hs37d5:19,940,511
INFO    2017-06-26 18:04:26 SortSam Finished reading inputs, merging and writing to output now.
 
Elapsed time: 525 secs
```

 | Young Collection Total Time | Old Collection Total Time | All Collections Total Time
 | ------------------------------------- | --------------------------------- | ----------------------------------
GC Count | 204 | 3 | 207
Average GC Time | 39 ms | 145 ms | 41 ms
Maximum GC Time | 94 ms | 205 ms | 205 ms
Total GC Time| 8 s | 437 ms | 8 s

Here is the file we used for our benchmarks: ftp://ftp.1000genomes.ebi.ac.uk/vol1/ftp/phase3/data/NA12878/exome_alignment/NA12878.mapped.ILLUMINA.bwa.CEU.exome.20121211.bam 
You are welcome to try it for you own benchmarks. Also could you provide information about your environment ? 
Here is our configuration: Ubuntu 16.04, i5-3570K, 3.40 GHz, 16 GB RAM, HDD
(what would you say about running some tests on a shared AWS instance to exclude the environment differences?)

2. We agree that this is potentially a  problem (though we haven't met it in our tests), but it is potentially a problem for any SortingCollection implementation (our implementation only multiplies the number of files by 3). 
And as soon the proposed changes should be turned ON explicitly, this side effect won't be a surprise.  We'll think about extending information in the javaDoc, and maybe in the readme. 

3. We have also thought and discussed such approaches a lot; as we see it, there is no ideal solution. 
For example, we foresee a problem with serial executor service: htsjdk already has some parallel feature implementation, AbstractAsyncWriter as example and approach this class uses can't be implemented by serial way, and all cases with using synchronization point such as BlockingQueue, Semaphore an so on will be restricted because it can't be implemented in serial way. 

4. We are fully ready to discuss a unified approach to parallelism in htsjdk, however we hope that this discussion won't block useful custom solutions (which could be rewritten once we get a unified approach). 

5. Agree, sorry about missing it, there is some test for AsyncWriteSortingCollection: 5eb9793bac7c864c65c30e314ac89bd2ee70be18